// lib/models/authentication-resource-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { User } from './user';
 /**
 * 
 *
 * @export
 * @interface AuthenticationResourceUser
 */
export interface AuthenticationResourceUser {

    /**
     * @type {User}
     * @memberof AuthenticationResourceUser
     */
    data: User;

    /**
     * @type {string}
     * @memberof AuthenticationResourceUser
     */
    token: string;
}