// lib/models/change-messaging-mode-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ChangeMessagingModeDto
 */
export interface ChangeMessagingModeDto {

    /**
     * @type {string}
     * @memberof ChangeMessagingModeDto
     */
    requesterId: string;

    /**
     * @type {string}
     * @memberof ChangeMessagingModeDto
     */
    interlocutorId: string;

    /**
     * @type {string}
     * @memberof ChangeMessagingModeDto
     */
    targetMode: ChangeMessagingModeDtoTargetModeEnum;

    /**
     * @type {string}
     * @memberof ChangeMessagingModeDto
     */
    projectId: string;
}

/**
 * @export
 * @enum {string}
 */
export enum ChangeMessagingModeDtoTargetModeEnum {
    OFF = 'OFF',
    LISTEN = 'LISTEN',
    ON = 'ON'
}